Backports by hand (hard work) ----------------------------------------------------------------------------------------- Fixes: Bursted Layout in screen "orderstatus" of ecommerce (OFBIZ-8265)
The layout of screen "order status" of application "ecommerce"  is busted .

I save the html of the screen into my local file, and It works fine. I don't know the reason exactly, but the patch attached by me works correctly.

jleroux: after review I agree it's the right way to fix it. OrderStatus.ftl indirectly calls OrderItems.ftl through the orderitems screen so nothing is lost when the addCommonToCartForm is moved from OrderStatus.ftl to OrderItems.ft, it's only a rearrangement which fixes the issue.

Thanks: fumiyoshi.to
-----------------------------------------------------------------------------------------

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release14.12@1761497 13f79535-47bb-0310-9956-ffa450edef68