"Applied fix from trunk for revision: 1761678" (tree conflict on OrderByReferrer.rptdesign handled by hand) ------------------------------------------------------------------------ r1761678 | jleroux | 2016-09-21 08:36:00 +0200 (mer. 21 sept. 2016) | 16 lignes
Fixes: Order By Referrer is not working
(OFBIZ-7318)

See
https://ofbiz-vm.apache.org:8443/ordermgr/control/OrderPurchaseReportOptions
Section : Order By Referrer

Found out that in OrderByReferrer.rptdesign we tried to fetch nonexisting fields
like "visitId" and "initialReferrer" from SalesOrderItemFact which causing 
problem in this report. On rev #886743 SalesOrderItemFact extended with visitId 
and initialReferrer and on rev #1055774 these field has been removed.
So I think we need to clean references of visitId and initialReferrer from the 
report and if we removed these fields then report will not make any sense. 
So in my opinion it is better to remove this report from OFBIZ.

Thanks: Rohit Koushalfor report and patch
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release15.12@1761680 13f79535-47bb-0310-9956-ffa450edef68