"Applied fix from trunk for revision: 1766074  " ------------------------------------------------------------------------ r1766074 | jleroux | 2016-10-21 19:22:29 +0200 (ven. 21 oct. 2016) | 15 lignes
Fixed: product page missing form elements
(OFBIZ-8489)

At r1746422, Deepak added a condition on EditProduct form. InventoryItemTypeId 
was enabled only for Product creation, else displayed as readonly, instead of 
editable dropdown

The only problem is that you can have the inventoryItemTypeId with a null value.
It's then considered by convention as not serialised. But then only the label 
shows. I fix it at by setting the inventoryItemTypeId as not serialised in this 
case. It was a bit tricky because you have to consider if the product is null 
(creation) and then not set product.inventoryItemTypeId. It would else defeat 
the creation case and give other contradictions.

Thanks: Wai for report
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release15.12@1766076 13f79535-47bb-0310-9956-ffa450edef68