"Applied fix from trunk for revision: 1779724" ------------------------------------------------------------------------ r1779724 | jleroux | 2017-01-21 12:15:17 +0100 (sam. 21 janv. 2017) | 26 lignes
Fixed: "View Order" throws exception in getReturnableQuantity
(OFBIZ-9157)

Steps to Reproduce:
- Create a sales order
- Add 1 orderitem with no unitPrice set (unitPrice = null) for the product
- set the status of the order to completed
- go to order-management and try to open the order for display like:
https://localhost:8443/ordermgr/control/orderview?orderId=XXXXX

Expected:
- the order is displayed with status completed

Actual:
- A groovy exception is displayed.

Analysis:
The service "getReturnableQuantity" throws an exception because it returns the 
returnablePrice = null, because the unitPrice of the orderItem was null.
The OUT field returnablePrice is mandatory according to the service-definition. 
The return parameter is set but the value is null, hence the exception.

Solution: adjust the service-definition to allow for "null" value 
of the returnablePrice OUT parameter.

Thanks: Leonard Lin
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1779728 13f79535-47bb-0310-9956-ffa450edef68