Fixed: Navigate Accounts tree does not work (OFBIZ-9191)
This is due to a change in (at least) Tomcat 8.0.39

Here is a simple fix. I tried other ways but none worked and I prefer to keep
this simple.
Using an encoder at this stage (just before calling he macro is not an option. 
Or maybe by using it on the linkurl variable alone. But then linkurl  must be 
transformed in a string and  back to a StringBuilder or more refactoring is 
needed.
Also using encode="true" for links in AccountingTrees.xml does not work.

This also removes trailing blanks (by hand)

Thanks: Moatasim Al Masri for report


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1781125 13f79535-47bb-0310-9956-ffa450edef68