Implemented: Load demo data shows error loading xml files Fixed: (OFBIZ-9211)
"gradlew cleanAll loadDefault" generates a data loading error

It's logically due to the removing of plugins/component-load.xml in OFBIZ-9182
A temporary solution is to put it in again. 
A definitive solution is to remove data dependencies. 
We have also this problem with ecommerce data which create dependencies 
from applications

jleroux: I don't close this issue yet because we need to work together on this.
I simply remove the regression by putting back plugins/component-load.xml

Thanks: Wai for report

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1782577 13f79535-47bb-0310-9956-ffa450edef68